requirements.test.txt
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-httpx==0.35.0
requests==2.32.3
tenacity==9.0.0